+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+- An issue with Pydantic typing, causing initialisation of `Benchmarker` to throw an
+  error.
+
+
 ## [v12.3.0] - 2024-03-13
 ### Changed
 - Updated `outlines` dependency to `>=0.0.36,<0.1`. This fixes a race condition caused